Create readme
# Overview
This code analyse data for emotional oddball task in children with severe mental health challenges. Overview of task, and population, and results from analysing a part of data can be found:

* 
* Electrophysiological correlates of fearful and sad distraction on target processing in adolescents with attention deficit-hyperactivity symptoms and affective disorders. 
* Singhal A, Shafer AT, Russell M, Gibson B, Wang L, Vohra S, Dolcos F. 
* Front Integr Neurosci. 2012 Dec 19;6:119. doi: 10.3389/fnint.2012.00119. PMID: 23267319; PMCID: PMC3525949.

## Task 
* Participants performed a modified version of the emotional oddball paradigm (Wang et al.,2005) which consisted of frequent stimuli serving as the baseline[scrambled pictures, 79% (465 trials)],infrequent distracters and odd ball targets, 21% (124 trials).Infrequent distracters consisted of sad and fearful pictures (13 trials each), neutral pictures (26 trials), and positive pictures (4 trials).The oddball targets(circles)were sub-grouped according to their preceding infrequent stimulus type [i.e.,target-after-sad (11 trials), target-after-fear (11 trials), target-after-target (24 trials), andtarget-after-neutral stimuli(22trials)].

* The participants’ task was to indicate whether the stimulus was a target or non-target by pressing a button. Participants were instructed to make a right hand button press any time they saw a target (circle) and a lef thand button press toall other stimuli (i.e., frequent scrambled and infrequent sad, fearful, neutral, and positive distracters). The inital , with targets and distractors. 

## Analysis 
This code extracts peaks and latency of ERP signals in a set of sites, whose electrode number and time window is specified in the code, and time series are given in 'C:\\QC\\'. 

As data is collected from children and adolescents with sever mental challenges, movement and noise of data was considerable. Subject's waveforms for AllDistractors
* AllDistractors = {'Fear','Neu','Sad','Standard'},

and AllTargets conditions  
* AllTargets = {'TargetafterFear','TargetafterNeu','TargetafterSad','TargetafterTarget'}

over Oz, O1, O2, T5, and T6 sites were first visually inpsected. Noisy Waveform were discarded from further analysis. The waveforms that survived this visual inspection are kept in "C:\good\" folder (sample for one subject is given).

Based on the survived participants' electrode/condition in the good data, peak and latency of ERP signals over the electrodes of interested were generated or discarded, based on the following association: 

- Good 'Oz' ==> generated 'Pz','Cz','PCz','POz','Oz'
- Good 'T5' ==> generated 'T5','PL','CL'
- Good 'T6' ==> generated 'T6','PR','CR'
- Good 'O1' ==> generated 'POL'
- Good 'O2' ==> generated 'POR'


- Good 'AllDistractors' ==> generated 'Fear','Neu','Sad','Standard'
- Good 'AllTargets'     ==> generated 'TargetafterFear','TargetafterNeu','TargetafterSad','TargetafterTarget'